coding repo

+++ nts.uk/uk.com/com.ctx/sys/portal/nts.uk.ctx.sys.portal.infra/src/main/java/nts/uk/ctx/sys/portal/infra/entity/layout/SptmtLayout.java
@@ -42,7 +42,7 @@ public class SptmtLayout extends UkJpaEntity implements Serializable, LayoutNew.
 
 	@Version
 	@Column(name = "EXCLUS_VER")
-	private BigDecimal exclusVer;
+	private int exclusVer;
 
 	@Column(name = "CONTRACT_CD")
 	private String contractCd;
